requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
fastapi
uvicorn[standard]
pymongo
python-jose[cryptography]
passlib[bcrypt]
